Clean up JCA smoke test resources
