+++ testsuite/integration/smoke/src/test/java/org/jboss/as/test/smoke/deployment/rar/tests/multiactivation/simple.xml
