+++ testsuite/integration/smoke/src/test/java/org/jboss/as/test/smoke/deployment/rar/tests/redeployment/ReDeploymentTestCase.java
@@ -19,104 +19,108 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.test.smoke.deployment.rar.tests;
+package org.jboss.as.test.smoke.deployment.rar.tests.redeployment;
 
-import java.io.IOException;
 import java.util.List;
 
-import javax.annotation.Resource;
+import javax.naming.Context;
 
+import org.jboss.arquillian.container.test.api.Deployer;
 import org.jboss.arquillian.container.test.api.Deployment;
+import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.junit.Arquillian;
+import org.jboss.arquillian.test.api.ArquillianResource;
+import org.jboss.as.arquillian.api.ContainerResource;
 import org.jboss.as.arquillian.api.ServerSetup;
 import org.jboss.as.arquillian.container.ManagementClient;
 import org.jboss.as.connector.subsystems.resourceadapters.Namespace;
-import org.jboss.as.connector.subsystems.resourceadapters.ResourceAdaptersExtension.ResourceAdapterSubsystemParser;
+import org.jboss.as.connector.subsystems.resourceadapters.ResourceAdaptersExtension;
 import org.jboss.as.test.integration.management.base.AbstractMgmtServerSetupTask;
-import org.jboss.as.test.integration.management.base.AbstractMgmtTestBase;
 import org.jboss.as.test.integration.management.base.ContainerResourceMgmtTestBase;
-import org.jboss.as.test.integration.management.util.MgmtOperationException;
 import org.jboss.as.test.shared.FileUtils;
 import org.jboss.as.test.smoke.deployment.rar.MultipleAdminObject1;
 import org.jboss.as.test.smoke.deployment.rar.MultipleConnectionFactory1;
 import org.jboss.dmr.ModelNode;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
-import org.jboss.shrinkwrap.api.asset.StringAsset;
 import org.jboss.shrinkwrap.api.spec.JavaArchive;
 import org.jboss.shrinkwrap.api.spec.ResourceAdapterArchive;
-import org.jboss.staxmapper.XMLElementReader;
-import org.jboss.staxmapper.XMLElementWriter;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import static org.junit.Assert.assertEquals;
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.OP;
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.OP_ADDR;
 import static org.junit.Assert.assertNotNull;
 
 
 /**
  * @author <a href="vrastsel@redhat.com">Vladimir Rastseluev</a>
- *         JBQA-5740 multiple resources deployment
+ *         JBQA-5968 test for undeployment and re-deployment
  */
 @RunWith(Arquillian.class)
-@ServerSetup(MultipleActivationTestCase.MultipleActivationTestCaseSetup.class)
-public class MultipleActivationTestCase extends ContainerResourceMgmtTestBase {
+@RunAsClient
+@ServerSetup(ReDeploymentTestCase.ReDeploymentTestCaseSetup.class)
+public class ReDeploymentTestCase extends ContainerResourceMgmtTestBase {
 
-    static class MultipleActivationTestCaseSetup extends AbstractMgmtServerSetupTask {
+    static String deploymentName = "re-deployment.rar";
+
+    @ContainerResource
+    private Context context;
+
+    @ArquillianResource
+    private Deployer deployer;
+
+    static class ReDeploymentTestCaseSetup extends AbstractMgmtServerSetupTask {
 
         @Override
-        public void doSetup(final ManagementClient managementClient) throws Exception {
-                String xml = FileUtils.readFile(MultipleActivationTestCase.class, "simple.xml");
-                List<ModelNode> operations = xmlToModelOperations(xml, Namespace.CURRENT.getUriString(), new ResourceAdapterSubsystemParser());
-                executeOperation(operationListToCompositeOperation(operations));
-        }
-
-        @Override
-        public void tearDown(final ManagementClient managementClient, final String containerId) throws IOException, MgmtOperationException {
-
+        public void tearDown(final ManagementClient managementClient, final String containerId) throws Exception{
             final ModelNode address = new ModelNode();
             address.add("subsystem", "resource-adapters");
-            address.add("resource-adapter", "archive.rar");
+            address.add("resource-adapter", deploymentName);
             address.protect();
             remove(address);
         }
+
+        @Override
+        protected void doSetup(final ManagementClient managementClient) throws Exception {
+
+        }
     }
 
+    private void setup() throws Exception {
+        String xml = FileUtils.readFile(ReDeploymentTestCase.class, "re-deployment.xml");
+        List<ModelNode> operations = xmlToModelOperations(xml, Namespace.CURRENT.getUriString(), new ResourceAdaptersExtension.ResourceAdapterSubsystemParser());
+        executeOperation(operationListToCompositeOperation(operations));
+
+        //since it is created after deployment it needs activation
+        final ModelNode address = new ModelNode();
+        address.add("subsystem", "resource-adapters");
+        address.add("resource-adapter", deploymentName);
+        address.protect();
+        final ModelNode operation = new ModelNode();
+        operation.get(OP).set("activate");
+        operation.get(OP_ADDR).set(address);
+        executeOperation(operation);
+
+    }
 
     /**
      * Define the deployment
      *
      * @return The deployment archive
      */
-    @Deployment
+    @Deployment(name = "re-deployment.rar", managed=false)
     public static ResourceAdapterArchive createDeployment() throws Exception {
-        String deploymentName = "archive.rar";
+
 
         ResourceAdapterArchive raa =
                 ShrinkWrap.create(ResourceAdapterArchive.class, deploymentName);
         JavaArchive ja = ShrinkWrap.create(JavaArchive.class, "multiple.jar");
-        ja.addPackage(MultipleConnectionFactory1.class.getPackage()).
-                addClasses(MultipleActivationTestCase.class, MgmtOperationException.class, XMLElementReader.class, XMLElementWriter.class);
-
-        ja.addPackage(AbstractMgmtTestBase.class.getPackage());
+        ja.addPackage(MultipleConnectionFactory1.class.getPackage());
         raa.addAsLibrary(ja);
 
-        raa.addAsManifestResource("rar/" + deploymentName + "/META-INF/ra.xml", "ra.xml")
-                .addAsManifestResource(new StringAsset("Dependencies: org.jboss.as.controller-client,org.jboss.dmr,org.jboss.as.cli\n"), "MANIFEST.MF");
-        ;
+        raa.addAsManifestResource(ReDeploymentTestCase.class.getPackage(), "ra.xml", "ra.xml");
         return raa;
     }
-
-    @Resource(mappedName = "java:jboss/name1")
-    private MultipleConnectionFactory1 connectionFactory1;
-
-    @Resource(mappedName = "java:jboss/name2")
-    private MultipleConnectionFactory1 connectionFactory2;
-
-    @Resource(mappedName = "java:jboss/Name3")
-    private MultipleAdminObject1 adminObject1;
-
-    @Resource(mappedName = "java:jboss/Name4")
-    private MultipleAdminObject1 adminObject2;
 
     /**
      * Test configuration
@@ -125,11 +129,13 @@ public class MultipleActivationTestCase extends ContainerResourceMgmtTestBase {
      */
     @Test
     public void testConfiguration() throws Throwable {
-        assertNotNull("CF1 not found", connectionFactory1);
-        assertNotNull("CF2 not found", connectionFactory2);
+        deployer.deploy(deploymentName);
+        setup();
+        deployer.undeploy(deploymentName);
+        deployer.deploy(deploymentName);
+        MultipleAdminObject1 adminObject1 = (MultipleAdminObject1) context.lookup("redeployed/Name3");
         assertNotNull("AO1 not found", adminObject1);
-        assertNotNull("AO2 not found", adminObject2);
-        assertEquals("not equal AOs", adminObject1, adminObject2);
+        deployer.undeploy(deploymentName);
     }
 
 }
