--- testsuite/integration/smoke/src/test/resources/rar/archive_ij1.rar/META-INF/ironjacamar.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ironjacamar>
-        <connection-definitions>
-        	<connection-definition  class-name="org.jboss.as.test.smoke.deployment.rar.MultipleManagedConnectionFactory2" jndi-name="java:jboss/D1CF2"  pool-name="D1CF2">    </connection-definition>
-        	<connection-definition  class-name="org.jboss.as.test.smoke.deployment.rar.MultipleManagedConnectionFactory1" jndi-name="java:jboss/D1CF1"  pool-name="D1CF2">    </connection-definition>
-		</connection-definitions>
-		 <admin-objects>
-          <admin-object class-name="org.jboss.as.test.smoke.deployment.rar.MultipleAdminObject1Impl"  jndi-name="java:jboss/D1AO1"	pool-name="D1AO1">  </admin-object>
-          <admin-object  class-name="org.jboss.as.test.smoke.deployment.rar.MultipleAdminObject2Impl" jndi-name="java:jboss/D1AO2" pool-name="D1AO2"> </admin-object>
-         </admin-objects>
-
-</ironjacamar>
